Standarize variables related to size

+++ src/Sandbox/World/ChunkManager.h
@@ -18,13 +18,13 @@ class ChunkManager
 	std::vector<glm::ivec3> _loadedChunksOrigin = {};
 	glm::ivec3 _lastChunkWithPlayer{};
 
-	int _renderDistance;
-	int _chunkSize;
-	unsigned _chunksToRender;
+	size_t _renderDistance;
+	size_t _chunkSize;
+	size_t _chunksToRender;
 	std::shared_ptr<WorldGenerator> _generator{};
 
 	glm::ivec3 GetNormalizedPosition(glm::vec3 position) const;
-	static unsigned CountChunksRecursive(unsigned level);
+	static size_t CountChunksRecursive(size_t level);
 	void RemoveExcludedChunks(const std::vector<glm::ivec3>& oldOrigins);
 
 	ChunkFrame CreateChunkFrame(const glm::ivec3& origin) const;
@@ -51,12 +51,12 @@ public:
 	void Bind(const std::shared_ptr<WorldGenerator>& worldGenerator);
 
 	/// @brief Get the amount of chunks that could be generated on the current render distance settings.
-	unsigned GetChunksToRenderCount() const;
+	size_t GetChunksToRenderCount() const;
 
 	/// @brief Get the camera object attached to this class.
 	Camera& GetCamera() const;
 
 	/// @brief Get the size of chunks that are generated.
-	unsigned GetChunkSize() const;
+	size_t GetChunkSize() const;
 };
